@@ -83,109 +86,104 @@
     "packaging": "jar"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "6.1.1",
-    "@fortawesome/free-solid-svg-icons": "6.1.1",
-    "@fortawesome/react-fontawesome": "0.1.18",
-    "@reduxjs/toolkit": "1.8.1",
-    "axios": "0.26.1",
-    "bootstrap": "5.1.3",
-    "dayjs": "1.11.0",
+    "@fortawesome/fontawesome-svg-core": "6.1.2",
+    "@fortawesome/free-solid-svg-icons": "6.1.2",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "@reduxjs/toolkit": "1.8.3",
+    "axios": "0.27.2",
+    "bootstrap": "5.2.0",
+    "dayjs": "1.11.4",
     "lodash": "4.17.21",
     "path-browserify": "1.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-hook-form": "7.29.0",
-    "react-jhipster": "0.20.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-hook-form": "7.30.0",
+    "react-jhipster": "0.21.0",
     "react-loadable": "5.5.0",
-    "react-redux": "7.2.8",
-    "react-redux-loading-bar": "5.0.2",
-    "react-router-dom": "5.3.0",
-    "react-toastify": "8.2.0",
-    "react-transition-group": "4.4.2",
-    "reactstrap": "9.0.1",
-    "redux": "4.1.2",
+    "react-redux": "8.0.2",
+    "react-redux-loading-bar": "5.0.4",
+    "react-router-dom": "6.3.0",
+    "react-toastify": "9.0.7",
+    "react-transition-group": "4.4.5",
+    "reactstrap": "9.1.2",
+    "redux": "4.2.0",
     "redux-thunk": "2.4.1",
-    "tslib": "2.3.1",
+    "tslib": "2.4.0",
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@testing-library/react": "12.1.4",
-    "@types/chai": "4.2.11",
-    "@types/jest": "27.4.1",
-    "@types/lodash": "4.14.181",
-    "@types/mocha": "7.0.2",
-    "@types/node": "16.11.26",
-    "@types/react": "17.0.43",
-    "@types/react-dom": "17.0.14",
-    "@types/react-redux": "7.1.23",
-    "@types/react-router-dom": "5.3.3",
+    "@testing-library/react": "13.3.0",
+    "@types/jest": "28.1.6",
+    "@types/lodash": "4.14.182",
+    "@types/node": "16.11.47",
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
+    "@types/react-redux": "7.1.24",
     "@types/redux": "3.6.31",
-    "@types/selenium-webdriver": "4.0.9",
-    "@types/webpack-env": "1.16.3",
-    "@typescript-eslint/eslint-plugin": "5.18.0",
-    "@typescript-eslint/parser": "5.18.0",
-    "autoprefixer": "10.4.4",
-    "browser-sync": "2.27.9",
+    "@types/webpack-env": "1.17.0",
+    "@typescript-eslint/eslint-plugin": "5.32.0",
+    "@typescript-eslint/parser": "5.32.0",
+    "autoprefixer": "10.4.8",
+    "browser-sync": "2.27.10",
     "browser-sync-webpack-plugin": "2.3.0",
-    "chai": "4.2.0",
-    "chai-as-promised": "7.1.1",
-    "concurrently": "7.1.0",
-    "copy-webpack-plugin": "10.2.4",
-    "core-js": "3.21.1",
+    "concurrently": "7.3.0",
+    "copy-webpack-plugin": "11.0.0",
+    "core-js": "3.24.1",
     "cross-env": "7.0.3",
     "css-loader": "6.7.1",
-    "css-minimizer-webpack-plugin": "3.4.1",
-    "eslint": "8.12.0",
+    "css-minimizer-webpack-plugin": "4.0.0",
+    "cypress": "10.4.0",
+    "cypress-audit": "1.1.0",
+    "eslint": "8.21.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-webpack-plugin": "3.1.1",
+    "eslint-plugin-cypress": "2.12.1",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-webpack-plugin": "3.2.0",
     "folder-hash": "4.0.2",
-    "fork-ts-checker-webpack-plugin": "7.2.3",
-    "generator-jhipster": "7.8.1",
+    "fork-ts-checker-webpack-plugin": "7.2.13",
+    "generator-jhipster": "7.9.1",
     "html-webpack-plugin": "5.5.0",
     "husky": "7.0.4",
     "identity-obj-proxy": "3.0.0",
-    "jest": "27.5.1",
-    "jest-junit": "13.1.0",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.3",
+    "jest-junit": "14.0.0",
     "jest-sonar-reporter": "2.0.0",
     "json-loader": "0.5.7",
-    "lint-staged": "12.3.7",
+    "lighthouse": "9.6.5",
+    "lint-staged": "13.0.3",
     "merge-jsons-webpack-plugin": "2.0.1",
-    "mini-css-extract-plugin": "2.6.0",
-    "mocha": "7.1.2",
-    "postcss-loader": "6.2.1",
-    "prettier": "2.6.2",
-    "prettier-plugin-java": "1.6.1",
-    "prettier-plugin-packagejson": "2.2.17",
-    "protractor": "7.0.0",
+    "mini-css-extract-plugin": "2.6.1",
+    "postcss-loader": "7.0.1",
+    "prettier": "2.7.1",
+    "prettier-plugin-java": "1.6.2",
+    "prettier-plugin-packagejson": "2.2.18",
     "react-infinite-scroll-component": "6.1.0",
     "redux-mock-store": "1.5.4",
     "rimraf": "3.0.2",
-    "sass": "1.49.11",
-    "sass-loader": "12.6.0",
+    "sass": "1.54.1",
+    "sass-loader": "13.0.2",
     "simple-progress-webpack-plugin": "2.0.0",
-    "sinon": "13.0.1",
-    "source-map-loader": "3.0.1",
+    "sinon": "14.0.0",
+    "source-map-loader": "4.0.0",
     "sourcemap-istanbul-instrumenter-loader": "0.2.0",
     "style-loader": "3.3.1",
-    "swagger-ui-dist": "4.10.3",
-    "terser-webpack-plugin": "5.3.1",
+    "swagger-ui-dist": "4.13.1",
+    "terser-webpack-plugin": "5.3.3",
     "thread-loader": "3.0.4",
-    "ts-jest": "27.1.4",
-    "ts-loader": "9.2.8",
-    "ts-node": "8.9.1",
-    "typescript": "4.6.3",
+    "ts-jest": "28.0.7",
+    "ts-loader": "9.3.1",
+    "typescript": "4.7.4",
     "wait-on": "6.0.1",
-    "webdriver-manager": "12.1.8",
-    "webpack": "5.71.0",
-    "webpack-cli": "4.9.2",
-    "webpack-dev-server": "4.8.0",
+    "webpack": "5.74.0",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "4.9.3",
     "webpack-merge": "5.8.0",
     "webpack-notifier": "1.15.0",
-    "workbox-webpack-plugin": "6.5.2"
+    "workbox-webpack-plugin": "6.5.4"
   },
   "engines": {
-    "node": ">=16.14.0"
+    "node": ">=16.16.0"
   },
   "cacheDirectories": [
     "node_modules"
