@@ -1,23 +1,23 @@
 import React from 'react';
-import { Switch } from 'react-router-dom';
+import { Route } from 'react-router-dom';
 
-import ErrorBoundaryRoute from 'app/shared/error/error-boundary-route';
+import ErrorBoundaryRoutes from 'app/shared/error/error-boundary-routes';
 
 import BankAccount from './bank-account';
 import BankAccountDetail from './bank-account-detail';
 import BankAccountUpdate from './bank-account-update';
 import BankAccountDeleteDialog from './bank-account-delete-dialog';
 
-const Routes = ({ match }) => (
-  <>
-    <Switch>
-      <ErrorBoundaryRoute exact path={`${match.url}/new`} component={BankAccountUpdate} />
-      <ErrorBoundaryRoute exact path={`${match.url}/:id/edit`} component={BankAccountUpdate} />
-      <ErrorBoundaryRoute exact path={`${match.url}/:id`} component={BankAccountDetail} />
-      <ErrorBoundaryRoute path={match.url} component={BankAccount} />
-    </Switch>
-    <ErrorBoundaryRoute exact path={`${match.url}/:id/delete`} component={BankAccountDeleteDialog} />
-  </>
+const BankAccountRoutes = () => (
+  <ErrorBoundaryRoutes>
+    <Route index element={<BankAccount />} />
+    <Route path="new" element={<BankAccountUpdate />} />
+    <Route path=":id">
+      <Route index element={<BankAccountDetail />} />
+      <Route path="edit" element={<BankAccountUpdate />} />
+      <Route path="delete" element={<BankAccountDeleteDialog />} />
+    </Route>
+  </ErrorBoundaryRoutes>
 );
 
-export default Routes;
+export default BankAccountRoutes;
