@@ -1,23 +1,23 @@
 import React from 'react';
-import { Switch } from 'react-router-dom';
+import { Route } from 'react-router-dom';
 
-import ErrorBoundaryRoute from 'app/shared/error/error-boundary-route';
+import ErrorBoundaryRoutes from 'app/shared/error/error-boundary-routes';
 
 import Operation from './operation';
 import OperationDetail from './operation-detail';
 import OperationUpdate from './operation-update';
 import OperationDeleteDialog from './operation-delete-dialog';
 
-const Routes = ({ match }) => (
-  <>
-    <Switch>
-      <ErrorBoundaryRoute exact path={`${match.url}/new`} component={OperationUpdate} />
-      <ErrorBoundaryRoute exact path={`${match.url}/:id/edit`} component={OperationUpdate} />
-      <ErrorBoundaryRoute exact path={`${match.url}/:id`} component={OperationDetail} />
-      <ErrorBoundaryRoute path={match.url} component={Operation} />
-    </Switch>
-    <ErrorBoundaryRoute exact path={`${match.url}/:id/delete`} component={OperationDeleteDialog} />
-  </>
+const OperationRoutes = () => (
+  <ErrorBoundaryRoutes>
+    <Route index element={<Operation />} />
+    <Route path="new" element={<OperationUpdate />} />
+    <Route path=":id">
+      <Route index element={<OperationDetail />} />
+      <Route path="edit" element={<OperationUpdate />} />
+      <Route path="delete" element={<OperationDeleteDialog />} />
+    </Route>
+  </ErrorBoundaryRoutes>
 );
 
-export default Routes;
+export default OperationRoutes;
