@@ -1,22 +1,23 @@
 import React from 'react';
-import { Switch } from 'react-router-dom';
-import ErrorBoundaryRoute from 'app/shared/error/error-boundary-route';
+import { Route } from 'react-router-dom';
+
+import ErrorBoundaryRoutes from 'app/shared/error/error-boundary-routes';
 
 import BankAccount from './bank-account';
 import Label from './label';
 import Operation from './operation';
 /* jhipster-needle-add-route-import - JHipster will add routes here */
 
-export default ({ match }) => {
+export default () => {
   return (
     <div>
-      <Switch>
+      <ErrorBoundaryRoutes>
         {/* prettier-ignore */}
-        <ErrorBoundaryRoute path={`${match.url}bank-account`} component={BankAccount} />
-        <ErrorBoundaryRoute path={`${match.url}label`} component={Label} />
-        <ErrorBoundaryRoute path={`${match.url}operation`} component={Operation} />
+        <Route path="bank-account/*" element={<BankAccount />} />
+        <Route path="label/*" element={<Label />} />
+        <Route path="operation/*" element={<Operation />} />
         {/* jhipster-needle-add-route-path - JHipster will add routes here */}
-      </Switch>
+      </ErrorBoundaryRoutes>
     </div>
   );
 };
