@@ -1,6 +1,7 @@
 import React from 'react';
 
-import ErrorBoundaryRoute from 'app/shared/error/error-boundary-route';
+import { Route } from 'react-router-dom';
+import ErrorBoundaryRoutes from 'app/shared/error/error-boundary-routes';
 import UserManagement from './user-management';
 import Logs from './logs/logs';
 import Health from './health/health';
@@ -8,15 +9,17 @@ import Metrics from './metrics/metrics';
 import Configuration from './configuration/configuration';
 import Docs from './docs/docs';
 
-const Routes = ({ match }) => (
+const AdministrationRoutes = () => (
   <div>
-    <ErrorBoundaryRoute path={`${match.url}/user-management`} component={UserManagement} />
-    <ErrorBoundaryRoute exact path={`${match.url}/health`} component={Health} />
-    <ErrorBoundaryRoute exact path={`${match.url}/metrics`} component={Metrics} />
-    <ErrorBoundaryRoute exact path={`${match.url}/configuration`} component={Configuration} />
-    <ErrorBoundaryRoute exact path={`${match.url}/logs`} component={Logs} />
-    <ErrorBoundaryRoute exact path={`${match.url}/docs`} component={Docs} />
+    <ErrorBoundaryRoutes>
+      <Route path="user-management/*" element={<UserManagement />} />
+      <Route path="health" element={<Health />} />
+      <Route path="metrics" element={<Metrics />} />
+      <Route path="configuration" element={<Configuration />} />
+      <Route path="logs" element={<Logs />} />
+      <Route path="docs" element={<Docs />} />
+    </ErrorBoundaryRoutes>
   </div>
 );
 
-export default Routes;
+export default AdministrationRoutes;
