@@ -1,16 +1,18 @@
 import React, { useState, useEffect } from 'react';
-import { Redirect, RouteComponentProps } from 'react-router-dom';
+import { Navigate, useLocation, useNavigate } from 'react-router-dom';
 
 import { useAppDispatch, useAppSelector } from 'app/config/store';
 import { login } from 'app/shared/reducers/authentication';
 import LoginModal from './login-modal';
 
-export const Login = (props: RouteComponentProps<any>) => {
+export const Login = () => {
   const dispatch = useAppDispatch();
   const isAuthenticated = useAppSelector(state => state.authentication.isAuthenticated);
   const loginError = useAppSelector(state => state.authentication.loginError);
   const showModalLogin = useAppSelector(state => state.authentication.showModalLogin);
   const [showModal, setShowModal] = useState(showModalLogin);
+  const navigate = useNavigate();
+  const location = useLocation();
 
   useEffect(() => {
     setShowModal(true);
@@ -20,13 +22,12 @@ export const Login = (props: RouteComponentProps<any>) => {
 
   const handleClose = () => {
     setShowModal(false);
-    props.history.push('/');
+    navigate('/');
   };
 
-  const { location } = props;
   const { from } = (location.state as any) || { from: { pathname: '/', search: location.search } };
   if (isAuthenticated) {
-    return <Redirect to={from} />;
+    return <Navigate to={from} replace />;
   }
   return <LoginModal showModal={showModal} handleLogin={handleLogin} handleClose={handleClose} loginError={loginError} />;
 };
