@@ -1,6 +1,5 @@
 import React from 'react';
-import { Router } from 'react-router-dom';
-import { createMemoryHistory } from 'history';
+import { MemoryRouter, Route, Routes } from 'react-router-dom';
 import { render } from '@testing-library/react';
 import { TranslatorContext } from 'react-jhipster';
 import configureStore from 'redux-mock-store';
@@ -8,7 +7,7 @@ import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 
 import { AUTHORITIES } from 'app/config/constants';
-import { PrivateRouteComponent, hasAnyAuthority } from './private-route';
+import PrivateRoute, { hasAnyAuthority } from './private-route';
 
 const TestComp = () => <div>Test</div>;
 
@@ -22,15 +21,19 @@ describe('private-route component', () => {
   const mockStore = configureStore([thunk]);
   const wrapper = (Elem: JSX.Element, authentication) => {
     const store = mockStore({ authentication });
-    return render(<Provider store={store}>{Elem}</Provider>);
+    return render(
+      <Provider store={store}>
+        <MemoryRouter>{Elem}</MemoryRouter>
+      </Provider>
+    );
   };
 
   // All tests will go here
-  it('Should throw error when no component is provided', () => {
+  it('Should throw error when falsy children are provided', () => {
     const originalError = console.error;
     console.error = jest.fn();
     expect(() =>
-      wrapper(<PrivateRouteComponent component={null} path="/" />, {
+      wrapper(<PrivateRoute>{null}</PrivateRoute>, {
         isAuthenticated: true,
         sessionHasBeenFetched: true,
         account: {
@@ -42,11 +45,10 @@ describe('private-route component', () => {
   });
 
   it('Should render an error message when the user has no authorities', () => {
-    const history = createMemoryHistory();
     const { container } = wrapper(
-      <Router history={history}>
-        <PrivateRouteComponent component={TestComp} path="/" />
-      </Router>,
+      <PrivateRoute>
+        <TestComp />
+      </PrivateRoute>,
       {
         isAuthenticated: true,
         sessionHasBeenFetched: true,
@@ -55,17 +57,14 @@ describe('private-route component', () => {
         },
       }
     );
-    expect(container.innerHTML).toEqual(
-      '<div class="insufficient-authority"><div class="alert alert-danger"><span>You are not authorized to access this page.</span></div></div>'
-    );
+    expect(container.innerHTML).toMatch(/<div class="insufficient-authority"><div class="alert alert-danger">.*<\/div><\/div>/);
   });
 
   it('Should render a route for the component provided when authenticated', () => {
-    const history = createMemoryHistory();
     const { container } = wrapper(
-      <Router history={history}>
-        <PrivateRouteComponent component={TestComp} path="/" />
-      </Router>,
+      <PrivateRoute>
+        <TestComp />
+      </PrivateRoute>,
       {
         isAuthenticated: true,
         sessionHasBeenFetched: true,
@@ -77,12 +76,19 @@ describe('private-route component', () => {
     expect(container.innerHTML).toEqual('<div>Test</div>');
   });
 
-  it('Should render a redirect to login when not authenticated', () => {
-    const history = createMemoryHistory();
+  it('Should redirect when not authenticated', () => {
     const { container } = wrapper(
-      <Router history={history}>
-        <PrivateRouteComponent exact component={TestComp} path="/" />
-      </Router>,
+      <Routes>
+        <Route
+          path="/"
+          element={
+            <PrivateRoute>
+              <TestComp />
+            </PrivateRoute>
+          }
+        />
+        <Route path="/login" element={<div>Login</div>} />
+      </Routes>,
       {
         isAuthenticated: false,
         sessionHasBeenFetched: true,
@@ -92,6 +98,7 @@ describe('private-route component', () => {
       }
     );
     expect(container.innerHTML).not.toEqual('<div>Test</div>');
+    expect(container.innerHTML).toEqual('<div>Login</div>');
   });
 });
 
