@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import javax.annotation.Generated;
 
-@Generated(value = "JHipster", comments = "Generated by JHipster 7.8.1")
+@Generated(value = "JHipster", comments = "Generated by JHipster 7.9.1")
 @Retention(RetentionPolicy.SOURCE)
 @Target({ ElementType.TYPE })
 public @interface GeneratedByJHipster {
