@@ -9,4 +9,8 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
  * See {@link tech.jhipster.config.JHipsterProperties} for a good example.
  */
 @ConfigurationProperties(prefix = "application", ignoreUnknownFields = false)
-public class ApplicationProperties {}
+public class ApplicationProperties {
+    // jhipster-needle-application-properties-property
+    // jhipster-needle-application-properties-property-getter
+    // jhipster-needle-application-properties-property-class
+}
