@@ -10,7 +10,10 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 /**
- * Spring Data SQL repository for the Operation entity.
+ * Spring Data JPA repository for the Operation entity.
+ *
+ * When extending this class, extend OperationRepositoryWithBagRelationships too.
+ * For more information refer to https://github.com/jhipster/generator-jhipster/issues/17990.
  */
 @Repository
 public interface OperationRepository extends OperationRepositoryWithBagRelationships, JpaRepository<Operation, Long> {
