@@ -2,8 +2,10 @@ package io.github.jhipster.sample.repository;
 
 import io.github.jhipster.sample.domain.Operation;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.IntStream;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import org.hibernate.annotations.QueryHints;
@@ -45,7 +47,9 @@ public class OperationRepositoryWithBagRelationshipsImpl implements OperationRep
     }
 
     List<Operation> fetchLabels(List<Operation> operations) {
-        return entityManager
+        HashMap<Object, Integer> order = new HashMap<>();
+        IntStream.range(0, operations.size()).forEach(index -> order.put(operations.get(index).getId(), index));
+        List<Operation> result = entityManager
             .createQuery(
                 "select distinct operation from Operation operation left join fetch operation.labels where operation in :operations",
                 Operation.class
@@ -53,5 +57,7 @@ public class OperationRepositoryWithBagRelationshipsImpl implements OperationRep
             .setParameter("operations", operations)
             .setHint(QueryHints.PASS_DISTINCT_THROUGH, false)
             .getResultList();
+        Collections.sort(result, (o1, o2) -> Integer.compare(order.get(o1.getId()), order.get(o2.getId())));
+        return result;
     }
 }
