@@ -3,7 +3,7 @@ import { translate } from 'react-jhipster';
 import { NavDropdown } from './menu-components';
 import EntitiesMenuItems from 'app/entities/menu';
 
-export const EntitiesMenu = props => (
+export const EntitiesMenu = () => (
   <NavDropdown
     icon="th-list"
     name={translate('global.menu.entities.main')}
