@@ -5,20 +5,21 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 
 export interface IMenuItem {
+  children: React.ReactNode;
   icon: IconProp;
   to: string;
   id?: string;
   'data-cy'?: string;
 }
 
-export default class MenuItem extends React.Component<IMenuItem> {
-  render() {
-    const { to, icon, id, children } = this.props;
+const MenuItem = (props: IMenuItem) => {
+  const { to, icon, id, children } = props;
 
-    return (
-      <DropdownItem tag={Link} to={to} id={id} data-cy={this.props['data-cy']}>
-        <FontAwesomeIcon icon={icon} fixedWidth /> {children}
-      </DropdownItem>
-    );
-  }
-}
+  return (
+    <DropdownItem tag={Link} to={to} id={id} data-cy={props['data-cy']}>
+      <FontAwesomeIcon icon={icon} fixedWidth /> {children}
+    </DropdownItem>
+  );
+};
+
+export default MenuItem;
